requirements.txt
streamlit
pandas
plotly
boto3
pymongo[srv]
python-dotenv
requests
numpy   